editor/backend/requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==2.2.12
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
paramiko==2.7.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
